chore: update firebase app check configuration
